+++ locale/lang-de.js
@@ -185,7 +185,7 @@ SnapTranslator.dict.de = {
     'translator_e-mail':
         'jens@moenig.org, jadga.huegle@sap.com', // optional
     'last_changed':
-        '2022-01-31', // this, too, will appear in the Translators tab
+        '2022-02-22', // this, too, will appear in the Translators tab
 
     // GUI
     // control bar:
@@ -863,6 +863,8 @@ SnapTranslator.dict.de = {
         'Index',
     'append %lists':
         'verbinde %lists',
+    'combinations %lists':
+        'Kombinationen %lists',
     'reshape %l to %nums':
         'strukturiere %l in %nums',
     'add %s to %l':