+++ src/threads.js
@@ -64,7 +64,7 @@ SnapExtensions, AlignmentMorph, TextMorph, Cloud, HatBlockMorph*/
 
 /*jshint esversion: 6*/
 
-modules.threads = '2022-February-09';
+modules.threads = '2022-February-26';
 
 var ThreadManager;
 var Process;
@@ -2348,70 +2348,93 @@ Process.prototype.doIfElse = function () {
 };
 
 Process.prototype.reportIfElse = function (block) {
-    var inputs = this.context.inputs;
+    var inputs = this.context.inputs,
+        accumulator,
+        condition,
+        expression,
+        trueIsBlock,
+        falseIsBlock;
 
     if (inputs.length < 1) {
+        // evaluate the first input, either a Boolean or a (nested) list
         this.evaluateNextInput(block);
-    } else if (inputs.length > 1) {
+        return;
+    }
+
+    if (inputs[0] instanceof List && this.enableHyperOps) {
+        // hyperize a (nested) list of Booleans
+        if (this.context.accumulator === null) {
+            // cache literal true/false cases for optimized evaluation
+            trueIsBlock = block.inputs()[1] instanceof BlockMorph;
+            falseIsBlock = block.inputs()[2] instanceof BlockMorph;
+            this.context.accumulator = {
+                results : [],
+                trueIsLiteral : !trueIsBlock,
+                trueCase : trueIsBlock ? null : block.inputs()[1].evaluate(),
+                falseIsLiteral : !falseIsBlock,
+                falseCase : falseIsBlock ? null : block.inputs()[2].evaluate()
+            };
+            // optimize if both true-/false- cases are literals
+            // for atomic conditions:
+            if (!trueIsBlock && !falseIsBlock) {
+                this.returnValueToParentContext(inputs[0].deepMap(
+                    leaf => leaf ? this.context.accumulator.trueCase
+                        : this.context.accumulator.falseCase)
+                );
+                this.popContext();
+                return;
+            }
+        } else if (inputs.length > 1) {
+            // retrieve & remember previous result & remove it from the inputs
+            this.context.accumulator.results.push(inputs.pop());
+        }
+        accumulator = this.context.accumulator;
+        if (accumulator.results.length === inputs[0].length()) {
+            // done with all the conditions in the current list
+            this.returnValueToParentContext(
+                new List(accumulator.results)
+            );
+            this.popContext();
+            return;
+        }
+        condition = inputs[0].at(accumulator.results.length + 1);
+
+        // optimize single literal true-/false- cases for atomic conditions:
+        if (!(condition instanceof List)) {
+            if (condition && accumulator.trueIsLiteral) {
+                accumulator.results.push(accumulator.trueCase);
+                return;
+            }
+            if (!condition && accumulator.falseIsLiteral) {
+                accumulator.results.push(accumulator.falseCase);
+                return;
+            }
+        }
+
+        this.pushContext(block); // recursive call
+        this.context.addInput(condition);
+        // optimize evaluation of literals:
+        this.context.accumulator = copy(accumulator);
+        this.context.accumulator.results = [];
+        return;
+    }
+
+    // handle a scalar condition
+    if (inputs.length > 1) {
+        // done with evaluating a case, retrieve and return its result
         if (this.flashContext()) {return; }
         this.returnValueToParentContext(inputs.pop());
         this.popContext();
+        return;
+    }
+    // this.assertType(inputs[0], ['Boolean']);
+    if (inputs[0]) {
+        expression = block.inputs()[1]; // true block
     } else {
-        // this.assertType(inputs[0], ['Boolean']);
-        if (inputs[0]) {
-            this.evaluateNextInput(block);
-        } else {
-            inputs.push(null);
-            this.evaluateNextInput(block);
-        }
+        expression = block.inputs()[2]; // false block
     }
+    this.pushContext(expression);
 };
-
-/*
-// Process - hyperized reporter-if, experimental, commented out for now
-
-Process.prototype.reportIfElse = function (block) {
-    var inputs = this.context.inputs;
-
-    if (inputs.length < 1) {
-        this.evaluateNextInput(block);
-    } else if (inputs.length > 1) {
-        if (this.flashContext()) {return; }
-        if (inputs[0] instanceof List && this.enableHyperOps) {
-            if (inputs.length < 3) {
-                this.evaluateNextInput(block);
-            } else {
-                this.returnValueToParentContext(
-                    this.hyperIf.apply(this, inputs)
-                );
-                this.popContext();
-            }
-        } else {
-            this.returnValueToParentContext(inputs.pop());
-            this.popContext();
-        }
-    } else {
-        if (inputs[0] instanceof List && this.enableHyperOps) {
-            this.evaluateNextInput(block);
-        } else {
-            // this.assertType(inputs[0], ['Boolean']);
-            if (inputs[0]) {
-                this.evaluateNextInput(block);
-            } else {
-                inputs.push(null);
-                this.evaluateNextInput(block);
-            }
-        }
-    }
-};
-
-Process.prototype.hyperIf = function (test, trueValue, falseValue) {
-    if (test instanceof List) {
-        return test.map(each => this.hyperIf(each, trueValue, falseValue));
-    }
-    return test ? trueValue : falseValue;
-};
-*/
 
 // Process process related primitives
 
@@ -3850,7 +3873,7 @@ Process.prototype.assertType = function (thing, typeString) {
     throw new Error(
         localize('expecting a') + ' ' +
         (typeString instanceof Array ?
-            typeString.reduce((a, b) => localize(a) + ' / ' + localize(b)) // +++
+            typeString.reduce((a, b) => localize(a) + ' / ' + localize(b))
             : localize(typeString)) +
         ' ' +
         localize('but getting a') + ' ' +
@@ -7660,8 +7683,8 @@ JSCompiler.prototype.compileInput = function (inp) {
         case 'Boolean':
             return '' + value;
         case 'text':
-            // enclose in double quotes
-            return '"' + value + '"';
+            // escape and enclose in double quotes
+            return '"' + this.escape(value) + '"';
         case 'list':
             return 'new List([' + this.compileInputs(value) + '])';
         default:
@@ -7682,7 +7705,7 @@ JSCompiler.prototype.compileInput = function (inp) {
         	}
          	// redirect var query to process
             return 'arguments[arguments.length - 1].getVarNamed("' +
-            	inp.blockSpec +
+            	this.escape(inp.blockSpec) +
             	'")';
         }
         return this.compileExpression(inp);
@@ -7693,4 +7716,48 @@ JSCompiler.prototype.compileInput = function (inp) {
             inp.constructor.name
         );
     }
+};
+
+JSCompiler.prototype.escape = function (string) {
+    var len = string.length,
+        result = '',
+        char,
+        esc,
+        i;
+    for (i = 0; i < len; i += 1) {
+        char = string[i];
+        switch (char) {
+        case '\\':
+            esc = '\\\\';
+            break;
+        case '\"':
+            esc = '\\"';
+            break;
+        case "\'":
+            esc = "\\'";
+            break;
+        case '\b':
+            esc = '\\b';
+            break;
+        case '\n':
+            esc = '\\n';
+            break;
+        case '\f':
+            esc = '\\f';
+            break;
+        case '\r':
+            esc = '\\r';
+            break;
+        case '\t':
+            esc = '\\t';
+            break;
+        case '\v':
+            esc = '\\v';
+            break;
+        default:
+            esc = char;
+        }
+        result += esc;
+    }
+    return result;
 };